Add env variable expansion

+++ src/shell.h
@@ -1,5 +1,6 @@
 #pragma once
 #include <stdbool.h>
+#include <stdint.h>
 #include <wait.h>
 
 #include "options.h"
@@ -7,7 +8,7 @@
 typedef struct Shell {
     pid_t pgid;
     pid_t fg_pgid;
-    int last_status;
+    uint8_t last_status;
     char cwd[MAX_PATHNAME_LENGTH];
     bool running;
 } Shell;
